pyproject.toml
[project]
name = "libspec"
version = "2.0.1"
description = "Spec-driven development library for LLM-assisted coding"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.13"
authors = [{ name = "Derek A. Rhodes", email = "physci@gmail.com" }]
keywords = ["specification", "llm", "codegen", "spec-driven", "development"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = ["docopt-ng>=0.9.0", "jinja2>=3.1.6", "lxml>=6.0.2", "mcp>=1.27.0", "xmldiff>=2.7.0"]

[project.urls]
Homepage = "https://github.com/drhodes/libspec"
Repository = "https://github.com/drhodes/libspec"
"Bug Tracker" = "https://github.com/drhodes/libspec/issues"


[project.scripts]
libspec = "libspec.cli:main"
libspec-mcp = "libspec.mcp_server:main"

[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv.workspace]
members = [
    ".", # the root (libspec package)
]

[dependency-groups]
dev = ["pudb>=2025.1.5", "pytest>=9.0.2"]
